Feature/initial system (#1)
* feat: Duplex and Transform streams

Worked on creating a message system with a Duplexx stream and sub-system message streams from
Transform streams. I believe a stream based system is the way to go, as backpressuring will allow
the system to load up large files over an internet connection, or from the file assets (out of scope
for the overall project, but good fore-thinking to put on the project roadmap). Transforms will
allow the sub systems to parse out the messages that pertain only to them, but will require some
playing to integrate into the overall system. This is a start, but I need to crack down and get to
testing to really see the benefits of this build. I believe another 2-3 hours of work will have this
system working and ready to begin integrating into the main project.

* feat: Message system tests and starting

Almost have the proof-of-concept working. Tests are passing, but I need to get the `node start`
command outputting properly. The function is not returning after await and I don't know why. This
could perhaps be a StackOverflow question

* feat(message system): Reading, Promises and awaiting

More in issue 65 of the sunset engine I'm working on

BREAKING CHANGE: Changed how the message system reads/writes

* feat(message system): The message system is complete and ready for implementaiton.

It is time to implement the message system. This stream based messaging system utilizes the
process.[stdin, stdout, stderr] streams built into node. While there could have been a basic event
based pub/sub scheme, this implementation will allow users to create streams for web hosts and
networking gameplay in the future.

* Apply suggestions from code review